examples/local-input-select-stt/requirements.txt
# Core dependencies
pipecat-ai[openai, silero]
textual==1.0.0
pydantic-settings==2.7.1
pyaudio==0.2.14
loguru
python-dotenv
numpy==1.26.4  # For audio processing
pyaudio
dotenv
openai